refactor buttons to use new naming scheme
