+++ frontend/src/components/ResultsView/ItemMenu/EventHeaderActions/EventHeaderActions.tsx
@@ -1,9 +1,9 @@
 import "./EventHeaderActions.css";
-import EditBtn from "../../../shared/buttons/EditBtn";
-import CloseBtn from "../../../shared/buttons/CloseBtn";
+import { EditBtn } from "../../../shared/buttons/editBtn.component";
+import { CloseBtn } from "../../../shared/buttons/closeBtn.component";
 import { useDispatch } from "react-redux";
 import { addPerson } from "../../../../redux/calculation/calculation-actions";
-import AddPersonBtn from "../../../shared/buttons/AddPersonBtn";
+import { AddPersonBtn } from "../../../shared/buttons/addPersonBtn.component";
 
 function EventHeaderActions({ editing, setEditing }: any) {
   const dispatch = useDispatch();
