+++ frontend/src/components/shared/buttons/deleteBtn.component.tsx
@@ -22,4 +22,4 @@ function DeleteBtn({
   );
 }
 
-export default DeleteBtn;
+export { DeleteBtn };
