+++ frontend/src/components/shared/buttons/resetBtn.component.tsx
@@ -81,4 +81,4 @@ function ResetBtn({ ariaTitle, className = "", iconClassName = "" }: any) {
   );
 }
 
-export default ResetBtn;
+export { ResetBtn };
